project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.10",
    "fastapi>=0.135.3",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "missingno>=0.5.2",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.13.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "typer>=0.24.1",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
